requirements.txt
# Azure Blob Storage SDK
azure-storage-blob

# AWS SDK for Python (Boto3)
boto3
